requirements.txt
Flask==2.3.3
easyocr==1.6.2
torch==2.3.1
numpy==1.26.4
Pillow==10.1.0
pillow-heif==0.9.2
pyheif==0.6.2
exifread==3.0.0
